refactor(docs): align editor page with playground format
- layout.tsx: HomeLayout wrapper (same as playground)
- page.tsx: dynamic import + loading spinner (SSR disabled)
- EditorClient.tsx: extracted from page.tsx, height adjusted for HomeLayout header
