+++ apps/docs/app/[lang]/editor/layout.tsx
@@ -1,0 +1,14 @@
+import { HomeLayout } from 'fumadocs-ui/layouts/home'
+import { baseOptions } from '@/lib/layout.shared'
+
+export default async function EditorLayout({
+  children,
+  params,
+}: {
+  children: React.ReactNode
+  params: Promise<{ lang: string }>
+}) {
+  const { lang } = await params
+
+  return <HomeLayout {...baseOptions(lang)}>{children}</HomeLayout>
+}
